Completed first integration with API. Not performant

+++ Funda.ProgrammingAssignment.Domain.Common/Dto/PropertyDto.cs
@@ -5,7 +5,7 @@
         public string Id { get; set; }
         public string Address { get; set; }
         public string City { get; set; }
-        public int RealEstateAgentId { get; set; }
+        public long RealEstateAgentId { get; set; }
         public string RealEstateAgentName { get; set; }
     }
 }
